feat: post fsd 구축
