+++ apps/web/src/features/Post/ui/Post/Post.tsx
@@ -1,12 +1,14 @@
 import React, { useEffect, useState, useRef } from "react";
 import "prismjs/themes/prism.css";
 import "prismjs/components/prism-javascript.min.js";
-import { getPostType } from "../../types/types";
-import { Block } from "../../types/BlockView/BlockView";
-import BlockView from "../../shared/BlockView/BlockView";
+
 import PostHeader from "./Header/Header";
 import Bars from "./Bars/Bars";
-import { PostValidation } from "../../services/Post/Post";
+
+import { getPostType } from "../../model/post/type";
+import { Block } from "../../../../shared/BlockView/model/BlockView/types";
+import BlockView from "../../../../shared/BlockView/ui/BlockView";
+import { PostValidation } from "../../libs/validation/Post";
 
 const PostView = (data: Partial<getPostType>) => {
   const contentRef = useRef<HTMLDivElement>(null);
