+++ apps/web/src/features/Post/ui/RecentPost/PostCard.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
-import { PostCardProps } from "../../types/RecentPost/RecentPost";
+import { PostCardProps } from "../../../../types/RecentPost/RecentPost";
 import { Clock, Heart, Eye } from "lucide-react";
 import { useSpring, animated } from "react-spring";
-import { formatRelativeTime } from "../../utils/times";
+import { formatRelativeTime } from "../../../../utils/times";
 
 const PostCard = ({ post, index, handlePost }: PostCardProps) => {
   const [isHovered, setIsHovered] = useState(false);
